Lookup Catalog Sources
* Save sources and metadata to parquet.
* Better output columns and formatting.

Observation Uploaded

* Use parquet loaded directly
