+++ lookup-catalog-sources/Dockerfile
@@ -2,7 +2,7 @@ ARG base_tag=latest
 
 FROM gcr.io/panoptes-exp/panoptes-utils:$base_tag
 
-RUN pip install google-cloud-pubsub
+RUN pip install google-cloud-pubsub pyarrow pendulum requests
 
 COPY . /app
 WORKDIR /app
