+++ lookup-catalog-sources/main.py
@@ -1,8 +1,11 @@
 import os
 import sys
-from io import StringIO
+from io import BytesIO
 
+import numpy as np
 import pandas as pd
+import pendulum
+import requests
 from astropy.utils.data import clear_download_cache
 from google.cloud import bigquery
 from google.cloud import firestore
@@ -22,6 +25,84 @@ BUCKET_NAME = os.getenv('BUCKET_NAME', 'panoptes-observations')
 
 PUBSUB_SUBSCRIPTION = 'read-lookup-catalog-sources'
 MAX_MESSAGES = os.getenv('MAX_MESSAGES', 1)
+
+FITS_HEADER_URL = 'https://us-central1-panoptes-exp.cloudfunctions.net/get-fits-header'
+
+SOURCE_COLUMNS = [
+    'picid',
+    'time',
+    'sequence_id',
+    'catalog_ra',
+    'catalog_dec',
+    'catalog_vmag',
+    'catalog_vmag_err',
+    'catalog_vmag_bin',
+    'x',
+    'y',
+    'x_int',
+    'y_int',
+    'twomass',
+    'gaia',
+    'unit_id',
+]
+
+METADATA_COLUMNS = {
+    'unit_id': 'unit_id',
+    'sequence_id': 'sequence_id',
+    'image_id': 'image_id',
+    'time': 'time',
+    'ra_mnt': 'mount_ra',
+    'ha_mnt': 'mount_ha',
+    'dec_mnt': 'mount_dec',
+    'ra_image': 'image_ra',
+    'dec_image': 'image_dec',
+    'airmass': 'image_airmass',
+    'moonsep': 'image_moonsep',
+    'moonfrac': 'image_moonfrac',
+    'exptime': 'image_exptime',
+    'camera_iso': 'camera_iso',
+    'camera_temp': 'camera_temp',
+    'camera_circconf': 'camera_circconf',
+    'camera_colortemp': 'camera_colortemp',
+    'camera_lens_serial_number': 'camera_lens_serial_number',
+    'camera_serial_number': 'camera_serial_number',
+    'camera_measured_ev': 'camera_measured_ev',
+    'camera_measured_ev2': 'camera_measured_ev2',
+    'camera_measured_r': 'camera_measured_r',
+    'camera_measured_g1': 'camera_measured_g1',
+    'camera_measured_g2': 'camera_measured_g2',
+    'camera_measured_b': 'camera_measured_b',
+    'camera_white_lvln': 'camera_white_lvln',
+    'camera_white_lvls': 'camera_white_lvls',
+    'camera_red_balance': 'camera_red_balance',
+    'camera_blue_balance': 'camera_blue_balance',
+    'site_latitude': 'site_latitude',
+    'site_longitude': 'site_longitude',
+    'site_elevation': 'site_elevation',
+    'bucket_path': 'bucket_path',
+    'public_url': 'public_url',
+}
+HEADER_COLUMNS = {
+    'IMAGEID': 'image_id',
+    'CRVAL1': 'image_ra_center',
+    'CRVAL2': 'image_dec_center',
+    'ISO': 'camera_iso',
+    'CAMTEMP': 'camera_temp',
+    'CIRCCONF': 'camera_circconf',
+    'COLORTMP': 'camera_colortemp',
+    'INTSN': 'camera_lens_serial_number',
+    'CAMSN': 'camera_serial_number',
+    'MEASEV': 'camera_measured_ev',
+    'MEASEV2': 'camera_measured_ev2',
+    'MEASRGGB': 'camera_measured_rggb',
+    'WHTLVLN': 'camera_white_lvln',
+    'WHTLVLS': 'camera_white_lvls',
+    'REDBAL': 'camera_red_balance',
+    'BLUEBAL': 'camera_blue_balance',
+    'LAT-OBS': 'site_latitude',
+    'LONG-OBS': 'site_longitude',
+    'ELEV-OBS': 'site_elevation',
+}
 
 # Storage
 try:
@@ -65,7 +146,6 @@ def process_topic(message):
     bucket_path = attributes.get('bucket_path')
 
     # Options
-    output_format = attributes.get('format', 'csv')
     force = attributes.get('force', False)
 
     # Get a document for the observation and the image. If only given a
@@ -133,18 +213,28 @@ def process_topic(message):
     catalog_sources['x_int'] = catalog_sources.x.astype(int)
     catalog_sources['y_int'] = catalog_sources.y.astype(int)
 
-    sources_bucket_path = f'{sequence_id}-sources.{output_format}'
+    # Get additional metadata.
+    unit_id, camera_id, observation_time = sequence_id.split('_')
+    catalog_sources['unit_id'] = unit_id
+    catalog_sources['sequence_id'] = sequence_id
+    catalog_sources['camera_id'] = camera_id
+    catalog_sources['time'] = pendulum.parse(observation_time).to_iso8601_string()
+    catalog_sources['catalog_vmag_bin'] = catalog_sources['catalog_vmag'].apply(np.floor).astype('int')
+
+    # Get just the columns we want.
+    catalog_sources = catalog_sources[SOURCE_COLUMNS]
+
+    sources_bucket_path = f'{sequence_id}-sources.parquet'
     logger.debug(f'Saving catalog sources to {sources_bucket_path}')
 
     # Lookup output format and save.
-    output_func = getattr(catalog_sources, f'to_{output_format}')
-    sio = StringIO()
-    output_func(sio, index=False)
-    sio.seek(0)
+    bio = BytesIO()
+    catalog_sources.to_parquet(bio, index=False)
+    bio.seek(0)
 
     # Upload
     obs_blob = output_bucket.blob(sources_bucket_path)
-    obs_blob.upload_from_file(sio)
+    obs_blob.upload_from_file(bio)
     logger.debug(f'Observation metadata saved to {obs_blob.public_url}')
 
     # Update observation status
@@ -174,14 +264,61 @@ def update_observation_file(sequence_id):
     metadata_df = pd.DataFrame([dict(image_id=doc.id, **doc.to_dict()) for doc in obs_query.stream()])
     print(f'{sequence_id}: Caching {len(metadata_df)} results from lookup')
 
-    sio = StringIO()
-    metadata_df.to_csv(sio, index=False)
-    sio.seek(0)
+    # Look up fits headers
+    fits_list = metadata_df.public_url.tolist()
+    print(f'Looking up headers for {len(fits_list)}')
+    headers = [get_header(f) for f in fits_list]
+    headers = [h for h in headers if type(h) == dict]
+
+    # Build DataFrame.
+    print(f'Making DataFrame for headers')
+    headers_df = pd.DataFrame(headers)[list(HEADER_COLUMNS.keys())].convert_dtypes().dropna()
+
+    # Clean values.
+    rggb_df = headers_df['MEASRGGB'].str.split(' ', expand=True).rename(columns={
+        0: 'camera_measured_r',
+        1: 'camera_measured_g1',
+        2: 'camera_measured_g2',
+        3: 'camera_measured_b',
+    }).astype('int')
+    headers_df = headers_df.join(rggb_df)
+    headers_df['CAMTEMP'] = headers_df.CAMTEMP.map(lambda x: int(x.split(' ')[0]))
+    headers_df['CIRCCONF'] = headers_df.CIRCCONF.map(lambda x: float(x.split(' ')[0]))
+
+    # Give better column names.
+    headers_df = headers_df.convert_dtypes().rename(columns=HEADER_COLUMNS)
+
+    print(f'Merging headers DataFrame with metadata')
+    metadata_df = metadata_df.merge(headers_df, on='image_id')
+
+    # Rename to friendlier columns.
+    metadata_df = metadata_df[list(METADATA_COLUMNS.keys())].rename(columns=METADATA_COLUMNS).convert_dtypes()
+
+    # Force dtypes on certain columns.
+    metadata_df['site_elevation'] = metadata_df['site_elevation'].astype('int')
+    metadata_df['camera_colortemp'] = metadata_df['camera_colortemp'].astype('int')
+    metadata_df['camera_lens_serial_number'] = metadata_df['camera_lens_serial_number'].astype('string')
+    metadata_df['camera_serial_number'] = metadata_df['camera_serial_number'].astype('string')
+
+    bio = BytesIO()
+    metadata_df.to_parquet(bio, index=False)
+    bio.seek(0)
 
     # Upload file object to public blob.
-    blob = output_bucket.blob(f'{sequence_id}-metadata.csv')
-    blob.upload_from_file(sio)
+    blob = output_bucket.blob(f'{sequence_id}-metadata.parquet')
+    blob.upload_from_file(bio)
     print(f'Observations file available at: {blob.public_url}')
+
+
+def get_header(bucket_path):
+    """Small helper function to lookup FITS header"""
+    res = requests.post(FITS_HEADER_URL, json=dict(bucket_path=bucket_path))
+    try:
+        h0 = res.json()['header']
+    except Exception:
+        h0 = dict()
+
+    return h0
 
 
 if __name__ == '__main__':
