+++ observation-file-uploaded/requirements.txt
@@ -1,4 +1,5 @@
 google-cloud-bigquery[pyarrow]
+google-cloud-storage
+numpy
 pandas
 pendulum
-requests
